Updating to use Fstrings
The only file with significant changes is the Lab1 Fundamentals notebook. Here, the section on formatted strings was updated by replacing the explanation of the format function with fstrings. There is still a small paragraph at the bottom of this section on the format function for reference as students may still come across it in older scripts.

In the rest of the Lab1 Fundamentals notebook and all other notebooks, changes were minor replacing the format syntax with fstrings whenever applicable.